crew_mcp_project/crew_mcp_project/requirements.txt
flask
crewai
crewai-tools
paramiko
python-dotenv
pyyaml
requests
fastmcp
ollama